add repository class
